requirements.txt
pymupdf
ollama
pdf2image
Pillow
gradio
requests
fastapi
uvicorn[standard]
python-multipart
